requirements.txt
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.29.4
beautifulsoup4==4.12.2
boto3==1.24.28
botocore==1.27.59
configparser==6.0.0
feedparser==6.0.10
IPython==8.12.0
nltk==3.8.1
numpy==1.24.3
openai==0.28.0
pandoc==2.3
pandas==2.1.1
pandera==0.17.2
pyarrow==13.0.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
PyPDF2==3.0.1
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==9.4.0
redivis==0.12.4
tenacity== 8.2.2
tiktoken==0.5.1
tqdm==4.64.0
matplotlib==3.7.1
Levenshtein==0.23.0
seaborn==0.13.0
tabulate==0.8.10